Redo the refactoring properly.
