+++ autonomous/__init__.py
@@ -1,2 +1,2 @@
-from autonomous.routines.basic_auto.square_auto import routine
+from autonomous.routines.square_auto.square_auto import routine
 from autonomous.utils.trajectory import CustomTrajectory
