+++ autonomous/routines/square_auto/square_auto.py
